output: generic simple tx json logger
Ticket: 3827
