--- src/output-json-bittorrent-dht.h
@@ -1,27 +1,0 @@
-/* Copyright (C) 2021 Open Information Security Foundation
- *
- * You can copy, redistribute or modify this Program under the terms of
- * the GNU General Public License version 2 as published by the Free
- * Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * version 2 along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
- * 02110-1301, USA.
- */
-
-/**
- * \file
- */
-
-#ifndef __OUTPUT_JSON_BITTORRENT_DHT_H__
-#define __OUTPUT_JSON_BITTORRENT_DHT_H__
-
-void JsonBitTorrentDHTLogRegister(void);
-
-#endif /* __OUTPUT_JSON_BITTORRENT_DHT_H__ */
