+++ src/output-json-ftp.c
@@ -154,60 +154,3 @@ bool EveFTPLogCommand(void *vtx, JsonBuilder *jb)
     jb_close(jb);
     return true;
 }
-
-
-static int JsonFTPLogger(ThreadVars *tv, void *thread_data,
-    const Packet *p, Flow *f, void *state, void *vtx, uint64_t tx_id)
-{
-    SCEnter();
-    OutputJsonThreadCtx *thread = thread_data;
-
-    const char *event_type;
-    if (f->alproto == ALPROTO_FTPDATA) {
-        event_type = "ftp_data";
-    } else {
-        event_type = "ftp";
-    }
-
-    JsonBuilder *jb =
-            CreateEveHeaderWithTxId(p, LOG_DIR_FLOW, event_type, NULL, tx_id, thread->ctx);
-    if (likely(jb)) {
-        if (f->alproto == ALPROTO_FTPDATA) {
-            if (!EveFTPDataAddMetadata(vtx, jb)) {
-                goto fail;
-            }
-        } else {
-            EveFTPLogCommand(vtx, jb);
-        }
-
-        OutputJsonBuilderBuffer(jb, thread);
-
-        jb_free(jb);
-    }
-    return TM_ECODE_OK;
-
-fail:
-    jb_free(jb);
-    return TM_ECODE_FAILED;
-}
-
-static OutputInitResult OutputFTPLogInitSub(ConfNode *conf,
-    OutputCtx *parent_ctx)
-{
-    AppLayerParserRegisterLogger(IPPROTO_TCP, ALPROTO_FTP);
-    AppLayerParserRegisterLogger(IPPROTO_TCP, ALPROTO_FTPDATA);
-    return OutputJsonLogInitSub(conf, parent_ctx);
-}
-
-void JsonFTPLogRegister(void)
-{
-    /* Register as an eve sub-module. */
-    OutputRegisterTxSubModule(LOGGER_JSON_TX, "eve-log", "JsonFTPLog", "eve-log.ftp",
-            OutputFTPLogInitSub, ALPROTO_FTP, JsonFTPLogger, JsonLogThreadInit, JsonLogThreadDeinit,
-            NULL);
-    OutputRegisterTxSubModule(LOGGER_JSON_TX, "eve-log", "JsonFTPLog", "eve-log.ftp",
-            OutputFTPLogInitSub, ALPROTO_FTPDATA, JsonFTPLogger, JsonLogThreadInit,
-            JsonLogThreadDeinit, NULL);
-
-    SCLogDebug("FTP JSON logger registered.");
-}
