--- src/output-json-tftp.h
@@ -1,29 +1,0 @@
-/* Copyright (C) 2017 Open Information Security Foundation
- *
- * You can copy, redistribute or modify this Program under the terms of
- * the GNU General Public License version 2 as published by the Free
- * Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * version 2 along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
- * 02110-1301, USA.
- */
-
-/**
- * \file
- *
- * \author Clément Galland <clement.galland@epita.fr>
- */
-
-#ifndef __OUTPUT_JSON_TFTP_H__
-#define __OUTPUT_JSON_TFTP_H__
-
-void JsonTFTPLogRegister(void);
-
-#endif /* __OUTPUT_JSON_TFTP_H__ */
